feat: Introduce singleton manifestations and signature management
This commit introduces a new concept of singleton manifestations, which are manifestations that can only have one item. This is useful for manuscripts, for example.

It also introduces a new `Signature` model for managing library sigla and signatures for each item, replacing the old `rism_id` field on the `Item` model. A new `Library` model has been added as well.

The forms and views have been updated to reflect the new models and logic, and new database migrations have been created.